requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==16.0.0
backports-abc==0.4
beautifulsoup4==4.5.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.23
certifi==2016.8.8
cffi==1.7.0
chardet==2.3.0
click==6.6
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
Django==1.10
django-filter==0.14.0
django-piston==0.2.3
djangorestframework==3.4.6
enum34==1.1.6
Flask==0.11.1
Flask-Login==0.3.2
genpac==1.4.1
gfwlist2pac==1.1.3
idna==2.1
image==1.5.4
ipaddress==1.0.16
itsdangerous==0.24
Jinja==1.2
Jinja2==2.8
jsonpath-rw==1.4.0
kombu==3.0.35
lxml==3.6.1
Markdown==2.6.6
MarkupSafe==0.23
mongoengine==0.10.6
MyProject==1.0
MySQL-python==1.2.5
numpy==1.11.1
objectpath==0.5
pandas==0.18.1
parsel==1.0.3
Pillow==3.3.1
ply==3.9
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycurl==7.43.0
PyDispatcher==2.0.5
Pygments==2.1.3
pymongo==3.3.0
pyOpenSSL==16.0.0
pypiwin32==219
pyquery==1.2.13
pyres==1.5
pyspider==0.3.7
python-dateutil==2.5.3
pytz==2016.6.1
queuelib==1.4.2
redis==2.10.5
requests==2.11.0
Scrapy==1.1.2
selenium==2.53.6
service-identity==16.0.0
setproctitle==1.1.10
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
Twisted==16.3.2
u-msgpack-python==2.1
virtualenv==15.0.3
w3lib==1.15.0
Werkzeug==0.11.10
WsgiDAV==1.2.0
zope.interface==4.2.0
